pyproject.toml
[project]
name = "pdf-highliter"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "akazdayo", email = "82073147+akazdayo@users.noreply.github.com" },
]
dependencies = [
    "spire-pdf>=9",
    "litellm>=1.65.4.post1",
    "google-genai>=1.9.0",
    "httpx>=0.28.1",
    "pypdf2>=3.0.1",
    "pypdf>=5.4.0",
    "pymupdf>=1.25.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.rye.scripts]
dev = { cmd = "python src/main.py", env-file = ".env" }

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_highliter"]
